[FIX] Broken editor handling for sys_language_uid -1
The given language does not exist within sites.
It also can not be resolved to a locale by default, further code and
fallback logic, maybe with a configuration option, needs to be added.

Therefore the current fix will catch the situation and prevent adding
the integration for editors.

That way the feature is gone, but elements can still be edited.

Resolves: #3